refactor(infra): one bounded close for the five that had written it
Five sites raced a close against a timer by hand: QueueConsumer's drain of a
worker that reached readiness before a later queue failed to, JobEvents, and all
three of @dunx/infra/amqp's closes. Two fixes had reached one copy each and
neither was carried across - unref on the timer, and the clearTimeout that stops
a resource which closed at once from holding the loop open for the rest of the
window. They now call closeWithin, which answers whether the bound expired and
leaves the warning to the caller.

@dunx/dashboard's bounded stays where it is. It bounds a read during a request
and answers with a value a panel renders, where these bound a close during
teardown and answer with a verdict; sharing them means either @dunx/dashboard
depending on @dunx/infra, which it does not, or a timeout primitive on
@dunx/core's public surface.

ResiliencePolicy stated a preference for AbortSignal.timeout that seven sites
contradicted. The rule that reconciles them: a signal where the operation takes
one, a race where it does not. None of these five closes takes a signal, so
there is nothing to hand one to. Written up in docs/architecture/message-brokers.md,
"Bounding a close", along with what cooperative cancellation would change and
where - the handler timeout, not the bound.

Closes #127
